# openapi.json
{
  "components": {
    "schemas": {
...
      "CredentialCreate": {
        "additionalProperties": false,
        "properties": {
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scheme_type": {
            "$ref": "#/components/schemas/CredentialScheme"
          },
          "secret": {
            "$ref": "#/components/schemas/CredentialSecretInput"
          }
        },
        "required": [
          "name",
          "scheme_type",
          "secret"
        ],
        "title": "CredentialCreate",
        "type": "object"
      },
      "CredentialRead": {
        "additionalProperties": false,
        "properties": {
          "configured": {
            "default": true,
            "title": "Configured",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "rotated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotated At"
          },
          "runtime_command_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Command Id"
          },
          "runtime_effect_state": {
            "$ref": "#/components/schemas/RuntimeEffectState"
          },
          "runtime_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Error Code"
          },
          "scheme_type": {
            "$ref": "#/components/schemas/CredentialScheme"
          }
        },
        "required": [
          "id",
          "project_id",
          "name",
          "scheme_type",
          "metadata",
          "created_by",
          "created_at",
          "rotated_at",
          "revoked_at",
          "runtime_effect_state",
          "runtime_command_id",
          "runtime_error_code"
        ],
        "title": "CredentialRead",
        "type": "object"
      },
      "CredentialRotate": {
        "additionalProperties": false,
        "properties": {
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "secret": {
            "$ref": "#/components/schemas/CredentialSecretInput"
          }
        },
        "required": [
          "secret"
        ],
        "title": "CredentialRotate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MCPAccessRead": {
        "additionalProperties": false,
        "properties": {
          "auth_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MCPAuthConfigRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "tokens": {
            "items": {
              "$ref": "#/components/schemas/MCPAccessTokenRead"
            },
            "title": "Tokens",
            "type": "array"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "auth_config",
          "tokens",
          "total",
          "page",
          "page_size"
        ],
        "title": "MCPAccessRead",
        "type": "object"
      },
      "MCPAccessStatusRead": {
        "additionalProperties": false,
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MCPAuthMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "remediation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation"
          },
          "runtime_command_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Command Id"
          },
          "runtime_effect_state": {
            "$ref": "#/components/schemas/RuntimeEffectState"
          },
          "runtime_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Error Code"
          }
        },
        "required": [
          "project_id",
          "mode",
          "configured",
          "remediation",
          "runtime_effect_state",
          "runtime_command_id",
          "runtime_error_code"
        ],
        "title": "MCPAccessStatusRead",
        "type": "object"
      },
      "MCPAccessTokenCreate": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "MCPAccessTokenCreate",
        "type": "object"
      },
...
      "MCPAuthConfigRead": {
        "properties": {
          "audiences": {
            "items": {
              "type": "string"
            },
            "title": "Audiences",
            "type": "array"
          },
          "issuer_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Url"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "mode": {
            "$ref": "#/components/schemas/MCPAuthMode"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "required_scopes": {
            "items": {
              "type": "string"
            },
            "title": "Required Scopes",
            "type": "array"
          },
          "runtime_command_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Command Id"
          },
          "runtime_effect_state": {
            "$ref": "#/components/schemas/RuntimeEffectState"
          },
          "runtime_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Error Code"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "format": "uuid",
            "title": "Updated By",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "mode",
          "issuer_url",
          "audiences",
          "required_scopes",
          "metadata",
          "updated_by",
          "updated_at",
          "runtime_effect_state",
          "runtime_command_id",
          "runtime_error_code"
        ],
        "title": "MCPAuthConfigRead",
        "type": "object"
      },
      "MCPAuthConfigUpdate": {
        "additionalProperties": false,
        "properties": {
          "allowed_algorithms": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Allowed Algorithms",
            "type": "array"
          },
          "audiences": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Audiences",
            "type": "array"
          },
          "issuer_url": {
            "anyOf": [
              {
                "format": "uri",
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Url"
          },
          "jwks_url": {
            "anyOf": [
              {
                "format": "uri",
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jwks Url"
          },
          "mode": {
            "$ref": "#/components/schemas/MCPAuthMode"
          },
          "required_scopes": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Required Scopes",
            "type": "array"
          }
        },
        "required": [
          "mode"
        ],
        "title": "MCPAuthConfigUpdate",
        "type": "object"
      },
...
      "Page_CredentialRead_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CredentialRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[CredentialRead]",
        "type": "object"
      },
...
      "RuntimeEffectState": {
        "enum": [
          "effective",
          "pending",
          "failed"
        ],
        "title": "RuntimeEffectState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted API-to-MCP builder/control plane",
    "title": "MCPlica API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/credentials": {
      "get": {
        "operationId": "list_credentials_api_v1_projects__project_id__credentials_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CredentialRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Credentials",
        "tags": [
          "credentials"
        ]
      },
      "post": {
        "operationId": "create_credential_api_v1_projects__project_id__credentials_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Credential",
        "tags": [
          "credentials"
        ]
      }
    },
    "/api/v1/projects/{project_id}/credentials/{credential_id}": {
      "delete": {
        "operationId": "revoke_credential_api_v1_projects__project_id__credentials__credential_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Credential",
        "tags": [
          "credentials"
        ]
      }
    },
    "/api/v1/projects/{project_id}/credentials/{credential_id}/rotate": {
      "post": {
        "operationId": "rotate_credential_api_v1_projects__project_id__credentials__credential_id__rotate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRotate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Credential",
        "tags": [
          "credentials"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/mcp-access": {
      "get": {
        "operationId": "get_mcp_access_api_v1_projects__project_id__mcp_access_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPAccessRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mcp Access",
        "tags": [
          "mcp-access"
        ]
      }
    },
    "/api/v1/projects/{project_id}/mcp-access/auth-mode": {
      "put": {
        "operationId": "update_auth_mode_api_v1_projects__project_id__mcp_access_auth_mode_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPAuthConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPAuthConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Auth Mode",
        "tags": [
          "mcp-access"
        ]
      }
    },
    "/api/v1/projects/{project_id}/mcp-access/status": {
      "get": {
        "operationId": "get_mcp_access_status_api_v1_projects__project_id__mcp_access_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPAccessStatusRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mcp Access Status",
        "tags": [
          "mcp-access"
        ]
      }
    },
    "/api/v1/projects/{project_id}/mcp-access/tokens": {
      "post": {
        "operationId": "create_access_token_api_v1_projects__project_id__mcp_access_tokens_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPAccessTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPAccessTokenIssued"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Access Token",
        "tags": [
          "mcp-access"
        ]
      }
    },
    "/api/v1/projects/{project_id}/mcp-access/tokens/{token_id}": {
      "delete": {
        "operationId": "revoke_access_token_api_v1_projects__project_id__mcp_access_tokens__token_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPAccessTokenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Access Token",
        "tags": [
          "mcp-access"
        ]
      }
    },
    "/api/v1/projects/{project_id}/mcp-access/tokens/{token_id}/rotate": {
      "post": {
        "operationId": "rotate_access_token_api_v1_projects__project_id__mcp_access_tokens__token_id__rotate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPAccessTokenRotate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPAccessTokenIssued"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Access Token",
        "tags": [
          "mcp-access"
        ]
      }
    },
...
  }
}